fix(log): 查询语句帮助入口改为 label 问号图标--story=163452967
