+++ src/dashboard-front/src/views/operate-data/access-log/components/SearchInput.vue
@@ -21,89 +21,6 @@
     ref="searchInputRef"
     class="search-input-container"
   >
-    <BkPopover
-      :is-show="searchUsage.showed"
-      trigger="click"
-      width="506"
-      theme="light"
-      placement="bottom"
-      ext-cls="access-log-popover"
-      @after-show="handlePopoverShow"
-      @after-hidden="handlePopoverHidden"
-    >
-      <div class="statement-example">
-        <AgIcon name="help-document" />
-        <span class="example-text">{{ t('如何使用') }}</span>
-      </div>
-      <template #content>
-        <div class="access-log-search-usage-content">
-          <div class="sample">
-            <div class="sample-item">
-              <p class="mode">
-                {{ t("关键字搜索") }}
-              </p>
-              <p class="value">
-                request_id: b3e2497532e54f518b3d1267fb67c83a
-                <AgIcon
-                  v-bk-tooltips="{ content: '引用条件' }"
-                  name="3-yuan-bohui"
-                  @click="() => handleClickUsageValue('request_id: b3e2497532e54f518b3d1267fb67c83a')"
-                />
-              </p>
-            </div>
-            <div class="sample-item">
-              <p class="mode">
-                {{ t("多个关键字匹配，支持 OR , AND") }}
-              </p>
-              <p class="value">
-                (app_code: "app-template" AND client_ip: "1.0.0.1") OR resource_name: get_user
-                <AgIcon
-                  v-bk-tooltips="{ content: '引用条件' }"
-                  name="3-yuan-bohui"
-                  @click="handleClickUsageValue('')"
-                />
-              </p>
-            </div>
-            <div class="sample-item">
-              <p class="mode">
-                {{ t("排除关键字") }}
-              </p>
-              <p class="value">
-                -status: 200
-                <AgIcon
-                  v-bk-tooltips="{ content: '引用条件' }"
-                  name="3-yuan-bohui"
-                  @click="() => handleClickUsageValue('-status: 200')"
-                />
-              </p>
-            </div>
-            <div class="sample-item">
-              <p class="mode">
-                {{ t("配置范围") }}
-              </p>
-              <p class="value">
-                request_duration: [5000 TO 30000]
-                <AgIcon
-                  v-bk-tooltips="{ content: '引用条件' }"
-                  name="3-yuan-bohui"
-                  @click="() => handleClickUsageValue('request_duration: [5000 TO 30000]')"
-                />
-              </p>
-            </div>
-          </div>
-          <div class="more">
-            {{ t("更多示例请参阅") }}
-            <a
-              class="link"
-              target="_blank"
-              :href="envStore.env.DOC_LINKS.QUERY_USE"
-            >
-              {{ t("“请求流水查询规则”") }}
-            </a>
-          </div>
-        </div>
-      </template>
-    </BkPopover>
     <BkDropdown
       :popover-options="popoverOptions"
       :disabled="queryHistory.length === 0"
@@ -138,8 +55,6 @@
 <script lang="ts" setup>
 // import { useGetGlobalProperties } from '@/hooks';
 import { useStorage } from '@vueuse/core';
-import AgIcon from '@/components/ag-icon/Index.vue';
-import { useEnv } from '@/stores';
 
 interface IProps {
   modeValue?: string
@@ -156,11 +71,9 @@ const {
 const emit = defineEmits<{
   'update:modeValue': [data: string]
   'search': [data: string]
-  'choose': [data: string]
 }>();
 
 const { t } = useI18n();
-const envStore = useEnv();
 // const globalProperties = useGetGlobalProperties();
 
 // 从本地存储获取搜索历史
@@ -176,7 +89,6 @@ const searchInputRef = ref(null);
 const localValue = ref('');
 const localPlaceholder = ref('');
 localPlaceholder.value = placeholder || t('请输入查询语句');
-const searchUsage = ref({ showed: false });
 
 watch(
   () => modeValue,
@@ -191,21 +103,6 @@ watch(
     emit('update:modeValue', payload);
   },
 );
-
-const handlePopoverShow = ({ isShow }: Record<string, boolean>) => {
-  searchUsage.value.showed = isShow;
-};
-
-const handlePopoverHidden = ({ isShow }: Record<string, boolean>) => {
-  searchUsage.value.showed = isShow;
-};
-
-const handleClickUsageValue = (innerText: string) => {
-  if (!innerText) {
-    innerText = '(app_code: "app-template" AND client_ip: "1.0.0.1") OR resource_name: get_user';
-  }
-  emit('choose', innerText);
-};
 
 const handleEnter = () => {
   emit('search', localValue.value);
@@ -225,70 +122,6 @@ defineExpose({ searchInputRef: searchInputRef.value });
     height: auto;
     border: none;
     border-radius: 0;
-  }
-}
-
-.search-input-container {
-  position: relative;
-
-  .statement-example {
-    position: absolute;
-    top: -26px;
-    right: 0;
-    display: flex;
-    font-size: 16px;
-    color: #3A84FF;
-    cursor: pointer;
-    align-items: center;
-
-    .example-text {
-      margin-left: 4px;
-      font-size: 12px;
-    }
-  }
-}
-
-.access-log-search-usage-content {
-  padding: 8px 4px -2px;
-
-  .sample {
-    padding-bottom: 12px;
-    margin-bottom: 8px;
-    border-bottom: 1px solid #DCDEE5;
-
-    .sample-item:not(:nth-last-child(1)) {
-      margin-bottom: 16px;
-    }
-
-    .mode {
-      margin-bottom: 4px;
-      font-family: MicrosoftYaHei-Bold;
-      font-size: 12px;
-      font-weight: bold;
-      color: #63656E;
-    }
-
-    .value {
-      font-size: 12px;
-      color: #63656E;
-
-      span {
-        margin-left: 2px;
-        font-size: 14px;
-        color: #3A84FF;
-        cursor: pointer;
-      }
-    }
-  }
-
-  .more {
-    font-size: 12px;
-    color: #63656e;
-
-    .link {
-      color: #3a84ff;
-      cursor: pointer;
-    }
   }
 }
 </style>
